+++ test/api/01_post_account.js
@@ -17,22 +17,21 @@ describe('post /account', () => {
     it('req', async () => {
         res = await req.post(path)
             .send({
-                email: 'root@mail.com',
+                email: 'god@mail.com',
                 password: 'banana',
-                role: 'root'
+                role: 'rw'
             })
         assert.equal(res.status, 201)
         dflow.verify('//trop/front/post_account_res#/body', res.body)
     })
 
-    // it('req => error', async () => {
-    //     res = await req.post(path)
-    //         .send({
-    //             email: 'banana@mail.com',
-    //             password: 'banana',
-    //             role: 'root'
-    //         })
-    //         console.log(res.body);
-    //     assert.equal(res.status, 409)
-    // })
+    it('req => error', async () => {
+        res = await req.post(path)
+            .send({
+                email: 'god@mail.com',
+                password: 'banana',
+                role: 'rw'
+            })
+        assert.equal(res.status, 409)
+    })
 })